Refactoring Engine (WIP).
