+++ tests/engine/Engine.test.js
@@ -1,38 +1,45 @@
-// import test from 'tape'
+import test from 'tape'
 
-// import Engine from '../../src/engine/Engine'
-// import FunctionManager from '../../src/function/FunctionManager'
-// import JsContext from '../../src/contexts/JsContext'
-// import MiniContext from '../../src/contexts/MiniContext'
-// import { libtestXML, libtest } from '../contexts/libtest'
+import Host from '../../src/host/Host'
+import Engine from '../../src/engine/Engine'
+import { wait } from '../testHelpers'
 
-// test('Engine: foo bar', t => {
-//   let { engine } = _setup()
-//   t.end()
-// })
+/*
+  Next:
+  - Find a concept for dealing with the async nature of the engine API
+*/
+test('Engine: simple cell', t => {
+  t.plan(1)
+  let { engine } = _setup()
+  // this should automatically trigger code analysis and
+  // incremental graph updates
+  engine.addCell({
+    id: 'sheet1.cell1',
+    lang: 'mini',
+    source: '1+2',
+    docId: 'sheet1'
+  })
+  // At any time the engine will have a number of pending actions
+  // such as updating the cell graph or triggering evaluations, like a job queue.
+  // Instead of processing the job queue automatically we could use a controllable
+  // scheduler, that we can trigger manually for testing.
+  // For real execution, we would apply an automated version, e.g. using a
+  // stop-and-go strategy, where the job queue is processed periodically (e.g. every 50ms).
+  wait(50)
+  .then(() => {
+    // TODO: in general it will not be necessary to
+    // wait for evaluations. For testing it would be nice,
+    // but we could also just wait for it.
+    return engine.step()
+  })
+  .then(() => {
+    debugger
+    t.fail('NOT IMPLEMENTED YET')
+  })
+})
 
-// function _setup() {
-//   let functionManager = new FunctionManager()
-//   functionManager.importLibrary('test', libtestXML)
-//   let contexts = {}
-//   let host = {
-//     createContext: function(language) {
-//       switch(language) {
-
-//       }
-//       if (language !== 'js') throw new Error('This stub only creates JsContexts')
-//       return Promise.resolve(jsContext)
-//     },
-//     functionManager
-//   }
-//   let jsContext = new JsContext()
-//   jsContext.importLibrary('test', libtest)
-//   let miniContext = new MiniContext(host)
-//   contexts['js'] = jsContext
-//   contexts['mini'] = miniContext
-//   let engine = new Engine(host)
-//   // HACK: don't let the engine run automatically by overriding the trigger
-//   engine._triggerScheduler = function() {}
-
-//   return { engine, host, contexts }
-// }
+function _setup() {
+  let host = new Host()
+  let engine = new Engine(host)
+  return { host, engine }
+}
